Updating configure/build scripts - Webplatform now optional.

+++ build/build/pack.js
@@ -101,14 +101,21 @@ module.exports = function (src, baton) {
     copyFolder(_c.CLIENTFILES, clientFilesDest);
     copyFolder(_c.DEPENDENCIES_BOOTSTRAP, bootstrapDest);
     copyFolder(_c.DEPENDENCIES_JNEXT, jnextDest);
-    copyFolder(_c.DEPENDENCIES_WEBPLATFORM_UI, _c.DEPLOY_UI);
 
     //Copy files to target directory (DO NOT copy webplatform-framework lib/* files over)
-    utils.copyFile(_c.DEPENDENCIES_WEBPLATFORM_FRAMEWORK_REQUIRE, browserRequireDest);
-    utils.copyFile(_c.DEPENDENCIES_WEBPLATFORM_FRAMEWORK_LIB, webplatformDest);
-    utils.copyFile(_c.DEPENDENCIES_WEBPLATFORM_I18N, webplatformDest);
     utils.copyFile(readmeFile, _c.DEPLOY);
     utils.copyFile(licenseFile, _c.DEPLOY);
+
+    //Copy webplatform files/folders if they exist
+    if (path.existsSync(_c.DEPENDENCIES_WEBPLATFORM)) {
+        copyFolder(_c.DEPENDENCIES_WEBPLATFORM_UI, _c.DEPLOY_UI);
+        utils.copyFile(_c.DEPENDENCIES_WEBPLATFORM_FRAMEWORK_REQUIRE, browserRequireDest);
+        utils.copyFile(_c.DEPENDENCIES_WEBPLATFORM_FRAMEWORK_LIB, webplatformDest);
+        utils.copyFile(_c.DEPENDENCIES_WEBPLATFORM_I18N, webplatformDest);
+    } else {
+        console.log("\n****WARNING: Webplatform could not be found.****");
+        console.log("****If building from the open-source community, please refer to the README for instructions on how to copy the Webplatform from a BB10 Webworks SDK installation.****\n");
+    }
 
     //Remove public folder
     wrench.rmdirSyncRecursive(_c.DEPLOY + 'lib/public', true);
